Update Athena.py
Things are a bit broke right now. Seems to be an issue with the Spotify API not accessing song list in current playlist, IMDB interactions are non-functional and the Kasa integrations have changed. 

Working on updating and refactoring

This version atleast will work, provided you generate your own API keys